Separate number screen into component
