+++ src/logic/calculate.js
@@ -1,4 +1,4 @@
-import operate from '../../operate';
+import operate from './operate';
 
 function isNumber(item) {
   return !!item.match(/[0-9]+/);
@@ -105,7 +105,7 @@ export default function calculate(obj, buttonName) {
     if (obj.total && !obj.next) {
       return { ...obj, operation: buttonName };
     }
-    
+
     if (!obj.total) {
       return { total: 0, operation: buttonName };
     }
